portfolio/requirements.txt
beautifulsoup4==4.12.3
icecream==2.1.3
python_dateutil==2.9.0.post0
python-binance==1.0.26
Requests==2.32.3
tabulate==0.9.0
